refactor(mappings): refactor update imports after changing the name of the mappings package to mapper (#11)
